refactor: restructure geometry structs
