+++ CMakeLists.txt
@@ -17,11 +17,11 @@ endif()
 find_package(ament_cmake REQUIRED)
 
 add_library(${PROJECT_NAME}
+  "src/geometry/point_2.cpp"
+  "src/geometry/point_3.cpp"
+  "src/geometry/transform2.cpp"
   "src/angle.cpp"
-  "src/number.cpp"
-  "src/point_2.cpp"
-  "src/point_3.cpp"
-  "src/transform2.cpp")
+  "src/number.cpp")
 
 target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -40,14 +40,14 @@ if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
 
   ament_add_gtest(${PROJECT_NAME}_tests
+    "test/geometry/point_2_test.cpp"
+    "test/geometry/point_3_test.cpp"
+    "test/geometry/transform2_test.cpp"
     "test/matrix/matrix_test.cpp"
     "test/matrix/square_matrix_test.cpp"
     "test/matrix/vector_test.cpp"
     "test/angle_test.cpp"
-    "test/number_test.cpp"
-    "test/point_2_test.cpp"
-    "test/point_3_test.cpp"
-    "test/transform2_test.cpp")
+    "test/number_test.cpp")
 
   target_include_directories(${PROJECT_NAME}_tests PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
