+++ include/keisan/geometry/point_2.hpp
@@ -18,10 +18,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef KEISAN__POINT_2_HPP_
-#define KEISAN__POINT_2_HPP_
+#ifndef KEISAN__GEOMETRY__POINT_2_HPP_
+#define KEISAN__GEOMETRY__POINT_2_HPP_
 
-#include <keisan/matrix/vector.hpp>
+#include "../matrix.hpp"
 
 namespace keisan
 {
@@ -66,4 +66,4 @@ struct Point2
 
 }  // namespace keisan
 
-#endif  // KEISAN__POINT_2_HPP_
+#endif  // KEISAN__GEOMETRY__POINT_2_HPP_
